cmd: Move stdiscosrv and strelaysrv in as subcommands to Syncthing
This adds `syncthing discosrv` and `syncthing relaysrv` as subcommands
to Syncthing. This has several advantages:

- If you have Syncthing, you also have these functions.
- It's fewer things for us to build and distribute, both for binaries,
  Debian archives and Docker images.
- We can take advantage of common functinality like auto upgrades
  (though that doesn't happen as part of this PR).

The main disadvantage is that it means we're carrying around more code
in the syncthing binary that might not be required for all users; the
increase on my system (darwin-arm64) is from 24.0 MB to 25.3 MB so I
think this is acceptable.

There is a slight change in the CLI handling for the commands: the Go
flag parser accepts long arguments with single dash, while kong does
not. I changed a few timeout values from integer to duration for
consistency, since it's anyway a breaking change. Given that nobody is
going to get hit by this by mistake (there are no auto upgrades from the
old binaries, and they need to run a different command) I think this is
fine, too.