gui, lib/api: Add possibility to feed through extra version information
This adds an environment variable STVERSIONEXTRA that, when set, gets
added to the version information in the API and GUI. In the CLI we see the following:

    % STVERSIONEXTRA=foo ./bin/syncthing --version
    syncthing v1.23.7-rc.1 (foo) "Fermium Flea" (go1.20.5 darwin...etc

In the GUI it's shown similarly in the About screen and in the local
device box.

The purpose of all this is to be able to communicate something about the
bundling or packaging, through the log & GUI and the end user, to the
potential person supporting it -- i.e., us. :) A wrapper can set this
variable to indicate that Syncthing is being run via `SyncTrayzor`,
`Syncthing-macOS`, etc., and thus indicate to the end user that the GUI
they are looking at is perhaps not the only source of truth and
management for this instance.