gui, lib/model: Disable periodic cleanup when versions are kept forever (fixes #8350)
Currently, the cleanup is set to run on a periodic timer regardless of
the days to keep versions. Because of this, even if versions are set to
be kept forever, the cleanup still runs on the timer. Furthermore, the
cleanup also runs even when using external versioning in spite of it
doing nothing for that versioning type.

This changes the behaviour as follows.

1. Cleanup is disabled when using external versioning.
2. Cleanup only runs when the days to keep versions are set to a value
   higher than "0".
3. When using Trash Can, Simple or Staggered versioning with the value
   above set to "0", the cleanup being "Disabled" is displayed in the
   folder information in the GUI even if the cleanup value is at its
   default.
4. In addition to the above, the cleanup input field is automatically
   disabled with a warning message informing the user about it being
   disabled when the number of days is set to "0".

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>