gui: Use case-insensive versions filter and allow backslashes as path separators (fixes #7973)
Currently, the versions filter is case-sensitive regardless of the
underlying OS. With this change, the filter becomes case-insensitive
everywhere, which is more user-friendly and makes it easier to search
for files whose exact case the user may not remember.

In addition, allow backslashes to be used as filter path separators.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>