model.go: use WARNING for "Unexpected folder" messages
So, this Syncthing client never complained that it could not sync a certain director^W folder to the Syncthing server. And with INFO messages ignored by monitoring, only logging in to the Syncthing GUI revealed that this new folder needed to be accepted by the server to be synced. As unexpected folder incidents may be uncommon enough, I feel WARNING to be more appropriate here.