gui: Remove Twitter link from footer
There is a Twitter link at the very bottom of the GUI. However, the page
seems rather dead, as almost no updates or information have been posted
there for a long time. On top of that, Twitter has recently switched to
a login wall, meaning that without logging in, people cannot even access
the site.

Because of the above, remove the Twitter link from the GUI altogether.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>