gui: Prevent modifications when saving changes (fixes #9019)
Currently, the GUI does not prevent the user from making further changes
even when it still has not saved the previous ones. This leads to the
newer changes being lost once the old ones are finally saved. The issue
is more apparent on slow hardware or when using the GUI remotely.

This change displays a modal that informs the user that the changes are
being saved and prevents the user from making any new modifications
until the saving process is complete.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>