gui: Disable Save button in modals when invalid or no changes
Currently, the modals have their Save buttons enabled regardless if
anything has been modified or the input values are valid or not. This
makes it so that the buttons become active only when the user has
either edited any of the values and the input values are valid.

For this to work in the Advanced Configuration modal, Folders in
advancedFolders, Devices in advancedDevices, and Default Folder, Device
and Ignore Patterns in advancedDefaults need not use separate forms for
each repeated element but rather they must be wrapped into a single,
larger form.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>