gui, lib/versioner: Allow to configure staggered versioning intervals (fixes #8352)
Currently, there are four intervals used in the staggered versioning,
with all of them but the Maximum Age being hard-coded. With this change,
the user is able to configure all intervals to their liking. Due to the
complexity of the configuration, this option is intended to be used by
advanced users only.

In addition to the above, also:

1. Add yet another interval for more fine-grained configuration
   possibilities. By default, the new interval activates after one year,
   keeping one version per month. However, because the default Maximum
   Age is also set to one year, this fifth interval is actually not
   utilised unless the user increases the Maximum Age to a higher value.

2. Fix a small bug in the Staggered Versioning test where the default
   time period for the third interval was supposed to be set to 30 days,
   but in reality it was set to just 7 days.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>