Move back percentage displays to HTML, for all needed cases.
Get rid of an extra <span> wrapping, instead declaring the ng-switch
attribute on the panel-status <div> itself.

Scanning and syncing information is now always shown properly in xs
and larger sizes.

The icon on the panel header is always visible, because there are
short spans where the percentage is not yet known / valid, which would
otherwise lead to a completely blank space instead of the progress
numbers.

Fix a syntax error in folderStatusText() function (missing semicolon).