Fix unwanted wrapping within the panel heading status.
In the device panels, a separate remote-devices-panel class was
introduced to wrap icons with a "display: inline-block" style.  It is
not needed for pure text nodes though, but on both icons.  However, it
only works when applied in an outer wrapping element around the icon
<span>.

Rename it to more generic inline-icon class and apply to both folder
and device status panels where needed.