lib/api: Improve ignore loading error handling (fixes #9253)
Any errors that were not a ParseError were returned as an API failure
rather than an error with the ignore file. This was quite limiting and
short sighted, since there may be permission errors, path format errors,
and possibly other things that are in fact legitimate problems with
loading the ignore file and not a bug or connection problem.

Returning a 500 error to the GUI causes the GUI to display a connection
error, which is never desirable here.