lib/model: Correct locking when removing folder runner (ref #9269)
A read lock is insufficient because the remove of the map entry happens
in the RemoveAndWait call, where previously it didn't.

However! I'm not sure this actually fixes the problem, because we then
release the lock and have a nil/unset runner for a while, which
previously would happen later under fmut again...