lib/api: Save session & CSRF tokens to database
This adds a "token manager" which handles storing and checking expired
tokens. It removes the old, corresponding CSRF manager which saved
things in a file. The result is less crap in the state directory, and
that active login sessions now survive a restart.

Best viewed with whitespace-insensitive diff, as a bunch of the auth
functions became methods instead of closures which changed indentation.