lib/protocol: handle empty names in unixOwnershipEqual (fixes #9039)
If syncOwnership is enabled and the remote uses for example a dockerized
Syncthing it can't fetch the ownername and groupname of the local
instance. Without this patch this led to an endless cycle of detected
changes on the remote and failing re-sync attempts.

This patch skips comparing the ownername and groupname if they are empty
on one side.

See https://github.com/syncthing/syncthing/issues/9039 for details.