gui, lib/scanner: Improve scan progress indication (ref #8331)
This adresses a few things that came up in issue #8331. It might fix the
issue, or it might not - it's not that clear-cut what exactly is
happening there.

This avoids sending progress updates when there is nothing to hash. And
ensures we send a final progress update when things are done. And in
that latter case, avoid showing 100% in the UI while the scan is
wrapping up.