gui: Show encryption status for devices sharing folder.
If a folder is shared to a device with encryption, add a lock icon in
front of the device name under "Shared With" in the folder details
panel.  Be careful not to add whitespace caused by line wraps in HTML
source code, which would defeat the purpose of keeping the icon glued
to the name by a non-breaking space.

Simplify redundant conditional comma code.

Apply the same HTML comment trick to avoid whitespace between a device
name (all but the last) and the separating comma.