gui: Add explanation to options enabled or disabled per folder type
Currently, some options are automatically enabled or disabled depending
on the folder type. However, there is no explanation in the GUI on why
the options are like that. Thus, add short explanatory notes to each
case, where the option is either disabled or enabled according to the
current folder type.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>