lib/api: Improve folder summary event, verbose service
This makes a couple of small improvements to the folder summary
mechanism:

- The folder summary includes the local and remote sequence numbers in
  clear text, rather than some odd sum that I'm not sure what it was
  intended to represent.
- The folder summary event is generated when appropriate regardless of
  whether there is an event listener. We did this before because
  generating it was expensive and we wanted to avoid doing it
  unnecessary. Nowadays, however, it's mostly just reading out
  pre-calcuated metadata, and anyway it's nice if it shows up reliably
  when running with -verbose
- A couple of fields that have been deprecated for years were removed
  ("version" instead of "sequence").

The point of all this is to make it easier to use these events to judge
when devices are in fact in sync. As-is, if I'm looking at two devices,
it's very difficult to reliably determine if they are in sync or not.
The reason is that while we can ask device A if it thinks it's in sync,
we can't see if the answer is "yes" because it has processed all changes
from B, or if it just doesn't know about the changes from B yet. With
proper sequence numbers in the event we can compare the two and
determine the truth. This makes testing a lot easier.