lib/fs,model: Clarify errors for Android filenames (fixes #9395)
Currently, Syncthing tries to sync files to Android devices with no
consideration for its filename restrictions (see [1]). This commit makes
it so that similarly to Windows, Syncthing relies on a hard-coded list
of characters which are invalid on Android and displays a specific error
message when user tries to sync a file with one of the characters in its
name.

[1] https://stackoverflow.com/a/64021421

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>