lib/model: Don't bump seq on error in index handler
This likely doesn't matter, as if the batch flush fails, the connection
likely failed. So we'll get a new connection with a new cluster config
and a new sequence to start from, thus it doesn't matter that we bump
the sequence even though we didn't actually successfully sent the
indexes. It's still wrong though :)