gui: Fix missing link to device editor for names with superscript (ref #9472) (#9494)
The commit 7e4e65ebf53dd98f0fee3f021c399b12f5966f12 added links to
devices listed in the Shared With list in the folder info. However, it
only added them to those that had no superscript next to them.

With this change, the links are added to all devices regardless of
whether they have the superscript next to their names or not. The commit
also simplifies the code by using anchors directly instead of wrapping
them in spans.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>